Refactor on import handlers
