+++ django_easy_drf/utils.py
@@ -1,0 +1,5 @@
+def get_viewset_name(model_class_name):
+    return f'{model_class_name}ViewSet'
+
+def get_serializer_name(model_class_name):
+    return f'{model_class_name}Serializer'
